Change AdvancedMarker's zIndex type to int

+++ packages/google_maps_flutter/google_maps_flutter_android/example/integration_test/google_maps_tests.dart
@@ -2240,7 +2240,7 @@ AdvancedMarker _copyAdvancedMarkerWithClusterManagerId(
     position: marker.position,
     rotation: marker.rotation,
     visible: marker.visible,
-    zIndex: marker.zIndex,
+    zIndex: marker.zIndex.toInt(),
     onTap: marker.onTap,
     onDragStart: marker.onDragStart,
     onDrag: marker.onDrag,
